fixed float64 showings
