+++ templates/metrix.page.tmpl
@@ -3,7 +3,7 @@
     <h1>Gauge Metrics</h1>
     <ul>
     {{range $key, $value := .GaugeMetrics}}
-        <li>{{$key}}: {{printf "%.09f" $value}}</li>
+        <li>{{$key}}: {{printf "%.6f" $value}}</li>
     {{end}}
     </ul>
     <h1>Counter Metrics</h1>
